fix: correct cors integration and remove duplicated security configuration
